refactor: remove dead code, extract shared validation
- Remove unused checkGhInstalled() and checkGhAuth() from github.js
- Remove dead writeDispatchContext() from dispatch-issue.js (superseded
  by writeIssueContext in dispatch-context.js)
- Extract validateOnboarded() into config.js to deduplicate shared
  onboarding validation from dispatch-issue.js and dispatch-pr.js
- Update dispatchIssue to use writeIssueContext from dispatch-context.js
