+++ lib/config.js
@@ -76,3 +76,17 @@ export function writeActive(data) {
   const content = yaml.dump(data);
   writeFileSync(activePath, content, 'utf8');
 }
+
+/**
+ * Validate that a repo is onboarded by checking projects.yaml.
+ * @param {string} repo - Repository in owner/repo format
+ * @throws {Error} If repo is not onboarded
+ */
+export function validateOnboarded(repo) {
+  const repoName = repo.split('/')[1];
+  const projects = readProjects();
+  const projectList = (projects && projects.projects) || [];
+  if (!projectList.find((p) => p.name === repoName)) {
+    throw new Error(`Repository "${repo}" is not onboarded. Run: rally onboard ${repo}`);
+  }
+}
