+++ lib/dispatch-issue.js
@@ -1,10 +1,11 @@
 import { execFileSync } from 'node:child_process';
-import { writeFileSync, mkdirSync, existsSync } from 'node:fs';
+import { existsSync } from 'node:fs';
 import { join, resolve } from 'node:path';
 import { createWorktree } from './worktree.js';
 import { createSymlink } from './symlink.js';
 import { addDispatch } from './active.js';
-import { readProjects } from './config.js';
+import { validateOnboarded } from './config.js';
+import { writeIssueContext } from './dispatch-context.js';
 import { launchCopilot } from './copilot.js';
 
 /**
@@ -19,37 +20,6 @@ export function slugify(title) {
     .replace(/[^a-z0-9]+/g, '-')
     .replace(/^-+|-+$/g, '')
     .slice(0, 50);
-}
-
-/**
- * Write a minimal dispatch-context.md into the worktree .squad/ directory.
- * Placeholder — will be superseded by lib/dispatch-context.js (#17).
- *
- * @param {string} squadDir - Path to .squad/ inside the worktree
- * @param {object} opts
- * @param {string} opts.repo - owner/repo
- * @param {number} opts.number - Issue number
- * @param {object} opts.issue - Issue object from GitHub
- */
-export function writeDispatchContext(squadDir, { repo, number, issue }) {
-  const labels = (issue.labels || []).map((l) => l.name || l).join(', ');
-  const assignees = (issue.assignees || []).map((a) => a.login || a).join(', ');
-  const content = `# Dispatch Context
-
-## Issue
-- **Repo:** ${repo}
-- **Issue:** #${number}
-- **Title:** ${issue.title}
-- **Labels:** ${labels || 'none'}
-- **Assignees:** ${assignees || 'none'}
-
-## Description
-${issue.body || '_No description provided._'}
-`;
-  if (!existsSync(squadDir)) {
-    mkdirSync(squadDir, { recursive: true });
-  }
-  writeFileSync(join(squadDir, 'dispatch-context.md'), content, 'utf8');
 }
 
 /**
@@ -99,12 +69,7 @@ export async function dispatchIssue(options = {}) {
   }
 
   // Validate repo is onboarded
-  const repoName = repo.split('/')[1];
-  const projects = readProjects();
-  const projectList = (projects && projects.projects) || [];
-  if (!projectList.find((p) => p.name === repoName)) {
-    throw new Error(`Repository "${repo}" is not onboarded. Run: rally onboard ${repo}`);
-  }
+  validateOnboarded(repo);
 
   const resolvedRepoPath = resolve(repoPath);
   const number = Number(issueNumber);
@@ -167,11 +132,7 @@ export async function dispatchIssue(options = {}) {
   }
 
   // 6. Write dispatch-context.md
-  const squadDir = join(worktreePath, '.squad');
-  if (!existsSync(squadDir)) {
-    mkdirSync(squadDir, { recursive: true });
-  }
-  writeDispatchContext(squadDir, { repo, number, issue });
+  writeIssueContext(worktreePath, { ...issue, number });
 
   // 7. Launch Copilot CLI
   const prompt = `Read .squad/dispatch-context.md and plan/implement a fix for issue #${number}`;
