+++ lib/dispatch-pr.js
@@ -4,7 +4,7 @@ import { join, resolve } from 'node:path';
 import { createWorktree } from './worktree.js';
 import { createSymlink } from './symlink.js';
 import { addDispatch } from './active.js';
-import { readProjects } from './config.js';
+import { validateOnboarded } from './config.js';
 import { writePrContext } from './dispatch-context.js';
 import { slugify } from './dispatch-issue.js';
 import { launchCopilot } from './copilot.js';
@@ -91,17 +91,13 @@ export async function dispatchPr(options = {}) {
   if (repoParts.length !== 2 || !repoParts[0] || !repoParts[1]) {
     throw new Error('Repository must be in "owner/repo" format');
   }
+  const repoName = repoParts[1];
   if (!repoPath) {
     throw new Error('Repository path is required');
   }
 
   // Validate repo is onboarded
-  const repoName = repoParts[1];
-  const projects = readProjects();
-  const projectList = (projects && projects.projects) || [];
-  if (!projectList.find((p) => p.name === repoName)) {
-    throw new Error(`Repository "${repo}" is not onboarded. Run: rally onboard ${repo}`);
-  }
+  validateOnboarded(repo);
 
   const resolvedRepoPath = resolve(repoPath);
   const number = Number(prNumber);
