+++ lib/github.js
@@ -1,34 +1,6 @@
 import { execFileSync } from 'node:child_process';
 
 /**
- * Checks if gh CLI is installed and available on PATH.
- * @throws {Error} If gh is not installed
- */
-export function checkGhInstalled() {
-  try {
-    execFileSync('gh', ['--version'], { encoding: 'utf8' });
-  } catch (error) {
-    throw new Error(
-      'GitHub CLI (gh) is not installed. Install from: https://cli.github.com/'
-    );
-  }
-}
-
-/**
- * Checks if gh CLI is authenticated.
- * @throws {Error} If gh is not authenticated
- */
-export function checkGhAuth() {
-  try {
-    execFileSync('gh', ['auth', 'status'], { encoding: 'utf8', stdio: 'pipe' });
-  } catch (error) {
-    throw new Error(
-      'GitHub CLI is not authenticated. Run: gh auth login'
-    );
-  }
-}
-
-/**
  * Gets issue details from GitHub.
  * @param {number|string} number - Issue number
  * @param {string} repo - Repository in owner/repo format
